+++ swift/Sources/main.swift
@@ -7,8 +7,8 @@ import JSONDecoding
 import KrakenAPI
 import SwiftCSV
 
-private let btcFormatter = createNumberFormatter(minimumFractionDigits: 8, maximumFranctionDigits: 8)
-private let fiatFormatter = createNumberFormatter(minimumFractionDigits: 2, maximumFranctionDigits: 2)
+let btcFormatter = createNumberFormatter(minimumFractionDigits: 8, maximumFranctionDigits: 8)
+let fiatFormatter = createNumberFormatter(minimumFractionDigits: 2, maximumFranctionDigits: 2)
 private let cryptoRateFormatter = createNumberFormatter(minimumFractionDigits: 0, maximumFranctionDigits: 6)
 private let fiatRateFormatter = createNumberFormatter(minimumFractionDigits: 0, maximumFranctionDigits: 2)
 
@@ -22,24 +22,6 @@ await storage.read()
 
 // Addresses that are part of the onchain wallet
 private let internalAddresses = Set<Address>(knownAddresses)
-
-func readCSVFiles(config: [(CSVReader, String)]) async throws -> [LedgerEntry] {
-    var entries = [LedgerEntry]()
-
-    try await withThrowingTaskGroup(of: [LedgerEntry].self) { group in
-        for (reader, filePath) in config {
-            group.addTask {
-                try await reader.read(filePath: filePath)
-            }
-        }
-
-        for try await fileEntries in group {
-            entries.append(contentsOf: fileEntries)
-        }
-    }
-
-    return entries
-}
 
 private func retrieveAndStoreTransactions(txIds: [String]) async -> [ElectrumTransaction] {
     let txIdsSet = Set<String>(txIds)
@@ -159,7 +141,7 @@ func round(_ value: Decimal, precision: Int, mode: NSDecimalNumber.RoundingMode 
 func readBtcAmount(_ amount: Double) -> Decimal {
     var decimalValue = Decimal(amount)
     var roundedValue = Decimal()
-    NSDecimalRound(&roundedValue, &decimalValue, 8, .down)
+    NSDecimalRound(&roundedValue, &decimalValue, 8, .plain)
 
     return roundedValue
 }
@@ -255,13 +237,6 @@ func electrumTransactionToLedgerEntries(_ transaction: ElectrumTransaction) asyn
     ) }
 }
 
-func formatAmount(_ entry: LedgerEntry) -> String {
-    let asset = entry.asset
-    let amount = entry.amount
-
-    return "\(asset.name) \(asset.type == .crypto ? btcFormatter.string(from: amount as NSNumber)! : fiatFormatter.string(from: amount as NSNumber)!)"
-}
-
 func formatRate(_ optionalRate: Decimal?, spendType: LedgerEntry.AssetType = .crypto) -> String {
     guard let rate = optionalRate else {
         return "unknown"
@@ -272,18 +247,6 @@ func formatRate(_ optionalRate: Decimal?, spendType: LedgerEntry.AssetType = .cr
     case .fiat: return fiatRateFormatter.string(from: rate as NSNumber)!
     }
 }
-
-struct Ref {
-    // "\(wallet)-\(id)"
-    let wallet: String
-    let id: String
-    let amount: Decimal
-    let rate: Decimal?
-}
-
-typealias RefsDeque = Deque<Ref>
-typealias RefsArray = [Ref]
-typealias Balance = [LedgerEntry.Asset: RefsDeque]
 
 private var ledgers = try await readCSVFiles(config: [
     (CoinbaseCSVReader(), "../data/Coinbase.csv"),
@@ -313,191 +276,24 @@ ledgers.sort(by: { a, b in a.date < b.date })
 //    print("\(entry.date) \(entry.wallet) \(entry.type) \(formatAmount(entry))")
 // }
 
-enum GroupedLedger {
-    // Single transaction within a wallet (e.g. Fee, Interest, Bonus) or ungrouped ledger entry
-    case single(entry: LedgerEntry)
-    // Trade within a single wallet
-    case trade(spend: LedgerEntry, receive: LedgerEntry)
-    // Transfer between wallets
-    case transfer(from: LedgerEntry, to: LedgerEntry)
-}
-
-let groupedLedgers: [GroupedLedger] = ledgers.reduce(into: [String: [LedgerEntry]]()) { groupIdToLedgers, entry in
-    switch entry.type {
-    // Group trades by ledger-provided groupId
-    case .trade:
-        groupIdToLedgers["\(entry.wallet)-\(entry.groupId)", default: [LedgerEntry]()].append(entry)
-    // Group deposit and withdrawals by amount (may lead to false positives)
-    case .deposit where entry.asset.type == .crypto,
-         .withdrawal where entry.asset.type == .crypto:
-        var id = "\(entry.asset.name)-\(btcFormatter.string(from: abs(entry.amount) as NSNumber)!)"
-
-        // Skip until we find a suitable group, greedy strategy
-        while groupIdToLedgers[id]?.count == 2 ||
-            groupIdToLedgers[id]?[0].type == entry.type
-        {
-            id += "-"
-        }
-
-        groupIdToLedgers[id, default: [LedgerEntry]()].append(entry)
-    default:
-        // Avoid grouping other ledger entries
-        groupIdToLedgers[UUID().uuidString] = [entry]
+let groupedLedgers: [GroupedLedger] = groupLedgers(ledgers: ledgers)
+let unmatchedTransfers = groupedLedgers.compactMap {
+    if case .single(let entry) = $0, entry.asset.type == .crypto, entry.type == .deposit || entry.type == .withdrawal {
+        return entry
     }
-
-    // let groupId = "\(entry.wallet)-\(entry.groupId)\(entry.type == .Fee ? "-fee" : "")"
-    // groupIdToLedgers[groupId, default: [LedgerEntry]()].append(entry)
-}.values.sorted { a, b in
-    a[0].date < b[0].date
-}.flatMap { group -> [GroupedLedger] in
-    switch group.count {
-    case 1: return [.single(entry: group[0])]
-    case 2 where group[0].type == .trade && group[0].amount > 0 && group[1].amount < 0:
-        return [.trade(spend: group[1], receive: group[0])]
-    case 2 where group[0].type == .trade && group[0].amount < 0 && group[1].amount > 0:
-        return [.trade(spend: group[0], receive: group[1])]
-    case 2 where group[0].type == .trade:
-        // Trade with 0 spend or receive, ungroup
-        return [.single(entry: group[0]), .single(entry: group[1])]
-    case 2 where group[0].type == .withdrawal && group[1].type == .deposit && group[0].wallet != group[1].wallet:
-        return [.transfer(from: group[0], to: group[1])]
-    case 2 where group[0].type == .deposit && group[1].type == .withdrawal && group[0].wallet != group[1].wallet:
-        return [.transfer(from: group[1], to: group[0])]
-    case 2 where group[0].type == group[1].type || group[0].wallet == group[1].wallet:
-        // Wrongly matched by amount, ungroup!
-        return [.single(entry: group[0]), .single(entry: group[1])]
-    default:
-        print(group)
-        fatalError("Group has more than 2 elements")
-    }
+    return nil
 }
 
-let BASE_ASSET = LedgerEntry.Asset(name: "EUR", type: .fiat)
-let BTC = LedgerEntry.Asset(name: "BTC", type: .crypto)
-
-//             [Wallet: Balance]
-var balances = [String: Balance]()
-for (index, group) in groupedLedgers.enumerated() {
-    print("Entry #\(index)")
-    switch group {
-    case .single(let entry):
-        print("\(entry.wallet) \(entry.type) \(formatAmount(entry)) - \(entry.id)")
-
-        // Not keeping track of base asset
-        guard entry.asset != BASE_ASSET else {
-            continue
-        }
-
-        var refs = balances[entry.wallet, default: Balance()][entry.asset, default: RefsDeque()]
-        if entry.amount > 0 {
-            refs.append(Ref(wallet: entry.wallet, id: entry.id, amount: entry.amount, rate: nil))
-        } else {
-            let removedRefs = subtract(refs: &refs, amount: -entry.amount)
-
-            if entry.type == .withdrawal {
-                let refsString = refs.map { "\($0.amount)@\(formatRate($0.rate, spendType: .fiat))" }.joined(separator: ", ")
-                let removedRefsString = removedRefs.map { "\($0.amount)@\(formatRate($0.rate, spendType: .fiat))" }.joined(separator: ", ")
-                print("  refs: \(removedRefsString)")
-                print("  balance: \(refsString)")
-            }
-        }
-        balances[entry.wallet, default: Balance()][entry.asset] = refs
-    case .transfer(let from, let to):
-        if from.wallet == to.wallet {
-            print("noop internal transfer \(from.wallet) \(formatAmount(to))")
-            continue
-        }
-        print("TRANSFER! \(from.wallet) -> \(to.wallet) \(formatAmount(to))")
-        guard var fromRefs = balances[from.wallet]?[from.asset] else {
-            fatalError("Transfer failed, balance is empty")
-        }
-
-        let subtractedRefs = subtract(refs: &fromRefs, amount: to.amount)
-        balances[from.wallet, default: Balance()][from.asset] = fromRefs
-        balances[to.wallet, default: Balance()][to.asset, default: RefsDeque()].append(contentsOf: subtractedRefs)
-        print("  Transfered refs:", subtractedRefs.map { "\($0.amount)@\(formatRate($0.rate, spendType: .fiat))" })
-    case .trade(let spend, let receive):
-        let wallet = spend.wallet
-        let rate = (-spend.amount / receive.amount)
-        print("\(wallet) trade! spent \(formatAmount(spend)), received \(formatAmount(receive)) @\(formatRate(rate, spendType: spend.asset.type))")
-        print("full spend \(spend.amount) rate \(rate) receive \(receive.amount)")
-
-        if spend.asset != BASE_ASSET {
-            // "move" refs to receive balance
-            var refs = balances[wallet, default: Balance()][spend.asset, default: RefsDeque()]
-            let balanceBefore = refs.reduce(0) { $0 + $1.amount }
-            print("  \(spend.asset.name) balance \(refs.reduce(0) { $0 + $1.amount })")
-            print("    bef: \(refs.map { $0.amount })")
-            let removedRefs = subtract(refs: &refs, amount: -spend.amount)
-            print("    rem: \(removedRefs.map { $0.amount })")
-            print("    aft: \(refs.map { $0.amount })")
-
-            let balanceAfter = (refs + removedRefs).reduce(0) { $0 + $1.amount }
-            if balanceBefore != balanceAfter {
-                fatalError("Balance subtract error, should be \(balanceBefore), it's \(balanceAfter)")
-            }
-
-            balances[wallet, default: Balance()][spend.asset] = refs
-
-            if receive.asset != BASE_ASSET {
-                let precision = receive.amount.significantFractionalDecimalDigits
-                // Propagate rate to receive side
-                // 🚨🚨 The operations here with the amount are not precise enough and leading to wrong balance
-                // TODO: receivedRefs total MUST match receive.amount
-                var receiveRefs = removedRefs.map {
-                    let nextRate = $0.rate.map { $0 * rate }
-                    let nextAmount = round($0.amount / rate, precision: precision)
-                    print("\(nextAmount) \($0.amount / rate)")
-                    // let nextAmount = $0.amount / rate
-
-                    return Ref(wallet: $0.wallet, id: $0.id, amount: nextAmount, rate: nextRate)
-                }
-                let dust = receive.amount - receiveRefs.sum
-                // TODO: alert if dust is significantly bigger than a rounding error
-                if dust != 0 {
-                    print("⚠️ Receive ref diff: \(dust), adding to first ref")
-                    guard let first = receiveRefs.first else {
-                        fatalError("Cannot fix rounding error, no elements")
-                    }
-
-                    receiveRefs[0] = Ref(
-                        wallet: first.wallet,
-                        id: first.id,
-                        amount: first.amount + dust,
-                        rate: first.rate
-                    )
-                }
-
-                let allReceiveRefs = balances[wallet, default: Balance()][receive.asset, default: RefsDeque()] + receiveRefs
-                balances[wallet, default: Balance()][receive.asset] = allReceiveRefs
-                print("  \(receive.asset.name) balance \(allReceiveRefs.reduce(0) { $0 + $1.amount })")
-                print("    \(allReceiveRefs.map { $0.amount })")
-                print("    \(receiveRefs.map { $0.amount })")
-
-                // receivedRefs total MUST match receive.amount or balances start to drift
-                if receiveRefs.sum != receive.amount {
-                    fatalError("Trade balance update error, should be \(receive.amount), is \(receiveRefs.reduce(0) { $0 + $1.amount })")
-                }
-            }
-
-            break
-        }
-
-        if receive.asset != BASE_ASSET {
-            // Add ref to balance
-            let ref = Ref(wallet: receive.wallet, id: receive.groupId, amount: receive.amount, rate: rate)
-            balances[receive.wallet, default: Balance()][receive.asset, default: RefsDeque()].append(ref)
-        }
-    }
+print("--- UNMATCHED TRANSFERS [\(unmatchedTransfers.count)] ---")
+for entry in unmatchedTransfers {
+    print(entry)
 }
 
-if let btcColdStorage = balances["❄️"]?[BTC] {
-    print("❄️")
-    print("total", btcColdStorage.sum)
-    print("total without rate", btcColdStorage.unknownSum)
-    // print("refs", btcColdStorage.description)
-}
-
-/**
- TODO: alert if dust is significantly bigger than a rounding error
- */
+// let balances = buildBalances(groupedLedgers: groupedLedgers)
+// let BTC = LedgerEntry.Asset(name: "BTC", type: .crypto)
+// if let btcColdStorage = balances["❄️"]?[BTC] {
+//    print("❄️")
+//    print("total", btcColdStorage.sum)
+//    print("total without rate", btcColdStorage.unknownSum)
+//    // print("refs", btcColdStorage.description)
+// }
